ok, so i found a usable slideshow online. i guess tomorrow i'll have to actually build the whole thing from scratch, because i am totally failing to center these divs right now!!! why isn't .wholething showing up at all???
and then later i'll have to hook up those images. this thing is so delicate! i tried hooking up internal urls and that didn't work...

oh yeah, and i have to hook up the actual functionality, but that should be doable, right? since this thing runs on react only, app.py lies forgotten, but i'll have to bring it back...