I thought about making a slideshow React app, but I already have one, so I deleted it and there are over 9000 changes detected. What I need to figure out, is why it's not displaying the image from the internal url.
I double slashed the backslashes, and it's still not working!

I should probably just put it in SQL instead of trying to deal with this images thing.

Oh great, now it thinks that my commit is empty. I mean... it's fine, as long as my local files stay the same, right? I think the only substantive change I made, is there was a picture of The Rolling Stones which I needed to convert from webp.

The point is, apparently I can make empty commits. That's pretty awesome!