I thought about deleting olds, and that was pointless because all I had to do was reinstall VS Code.(I did end up putting the olds in cool named folders though! Ooh!)
Reinstalling was startlingly easy, like, I didn't have to re hook it up with Github or anything. It's like, did I actually just reinstall it??? I mean, that was really convenient. Let me try gitignoring my .env, that's the real question.

That worked, but that .env is like, a weird remnant of my slideshow app I thought about making? I think? Anyway, .gitignore is an actual file that exists in the folder, remember that!

Anyway, I was trying to get slideshow.py working, I tried encoding and actually using the filename. Now look, I'm getting the same error, even though I commented most of the code out!!! What is going on???

Oh, so maybe all that stuff was dependent on Flask. Because I put the flask stuff in, and I got a different error.

OK, I tried using get, which is a crazy to me, because I didn't have to use it up in server! But down here, it opens up a whole new can of errors, sooo.... um, it says that it's using the password NO. Is that just from the default stuff it puts in from get? I guess I have to figure out why the original isn't working...