Aha! The issue was pymysql, I'm sure I'm not formatting that correctly. See, I finally gave up and asked ChatCPT, and, well, see, ChatGPT really is helpful...
Great, I changed it a little (used more backquotes...) but that didn't fix it! See, I'm not sure if it likes my f string. Not sure how else to do it?

Oh, well according stackoverflow users, I shouldn't be forming query strings with string formatting at all because it exposes me to injection attacks. Well, how am I supposed to switch to placeholders??? Oh, is it these '%s' things? That's kind of similar to that (sql, (user)) formatting.

OK, so that was an easy fix! Now...

Now I've got an error that's kind of similar looking, but it is different.
 query = query % self._escape_args(args, conn)
TypeError: unsupported operand type(s) for %: 'tuple' and 'str'

Whoa whoa whoa! Thanks ChatGPT! Apparently I can separate my params! Hehehe, I sure hope ChatGPT isn't cheating, because it is nice. Right, so the question is, why was (user) my params last time? I really don't get that! I will actually have to ask about that!

Well... it's almost working... for some reason, it seems to be reading my img_data starting at 1, when I'm telling it to start at 0!!! Not sure what the deal is with that!!! No, it appears to be printing... 2 lines? It appears to be printing 2 lines at a time for each of these. ??? I could probably ask ChatGPT about it, but I will try sleeping on it instead! Sleep still beats ChatGPT!