Maybe this isn't the MOST BEST way to do it, but with a set of data so limited, I think I could pull it off with just separating read_data from fill_database, and then processing the data separately in both fill_database and fetch_data.
There, OK, putting the images in React, I might need JSON & Axios for that...

But first, what I don't get, is how to change the state!!! And what's the deal with onLoginChange?

See, at the top of index.js, I'm trying to import a variable, like in Python. I guess this is where I'll need to use JSON!!!

I'd better save this before I try anything...