Something wrong with my parsing! When I tried to 'var obj = JSON.parse(glossary)', it went 'SyntaxError: "[object Object]" is not valid JSON'.
However, I just commented that parsing line, and it's still having that error, so maybe the parsing was not the problem.

And I've also seen it complaining about 'Unexpected token '<', "<!DOCTYPE "... is not valid JSON', and I don't know what it's talking about! I don't see any DOCTYPEs around here!!!