I mean, TECHNICALLY I could include create_table in server.py, kind of like the lesson has resetdb in there. Is this helping anything??? Probably not!!! I'm just trying to find something to do!!!
And I have create_table operating on filename here, and I'm sure I have too many things operating on filename!!! In the lesson, it's dealing with counts too. I guess if I REALLY wanted to make this similar to the lesson, I'd have like, 3 variables in play... I'll try that, although it seems like an overcomplication at this point. Well, I'll try it, and see if ANY of it works...