ok, maybe i'm overcomplicating this. i should be creating my address variable in python, right? so i'm trying simple concatenation but it doesn't seem to be working. not to mention i might be transferring the variable over wrong too. do i need to add it to the list of variables first?
i'm thinking of going back and finding all the unecessary things and deleting them to make it less confusing, but let me save this first.

wait wait, here's a promising looking answer with string concatenation in js. i'm going to try that, but here's by failed attempt at making an address in python.